backend/app/generate_hash.py
from passlib.context import CryptContext

pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

password = "Admin@123"
hashed = pwd_context.hash(password)

print("Hashed password:", hashed)
